Update 1._Community_Specification_License_v0.9.md
I think the license text itself should be licensed under a "ND" type license (though we can certainly discuss). In listing the license for the text, the question of who 'owns' the community spec license text needs to be answered and so to start of this discussion I've proposed JDF but am interested in the group's thoughts here. - Scott